backend/pyproject.toml
[project]
name = "avatar-ai-backend"
version = "0.1.0"
description = "Avatar AI Demo backend — FastAPI + Pipecat"
requires-python = ">=3.12"

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pyyaml>=6.0.2",
    "anthropic>=0.39.0",
    "python-multipart>=0.0.12",
    "sse-starlette>=2.1.3",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py312"
